Implémentation de la supperssion d'un livre

+++ controllers/book.js
@@ -1,5 +1,6 @@
+const { error } = require("console");
 const Book = require("../models/Book");
-
+const fs = require ("fs");
 exports.addBook = (req, resp, next) => {
     const bookObject = JSON.parse(req.body.book);
     delete bookObject._id;
@@ -26,7 +27,7 @@ exports.updateBook = (req, resp, next) => {
 
     Book.findOne({_id: req.params.id})
     .then((book) => {
-        book.userId != req.auth.userId ? resp.status(401).json({message : "No authorized !!!"}) : Book.updateOne({_id: req.params.id}, {...bookObject, _id: req.params.id})
+        book.userId != req.auth.userId ? resp.status(401).json({message : "unauthorized !!!"}) : Book.updateOne({_id: req.params.id}, {...bookObject, _id: req.params.id})
         .then(() => resp.status(200).json({message: "book successfully updated !"}))
         .catch(error => resp.status(401).json({error}))
     })
@@ -34,9 +35,20 @@ exports.updateBook = (req, resp, next) => {
  };
 
 exports.deleteBook = (req,resp,next) => {
-    Book.deleteOne({_id: req.params.id})
-    .then(() => resp.status(200).json({ message : "Book successfully deleted"}))
-    .catch(error => resp.status(400).json({ error }))
+    Book.findOne({_id: req.params.id})
+    .then(book => {
+        if(book.userId != req.auth.userId) {
+            resp.status(401).json({message: "unauthorized!!!"})
+        } else {
+            const filename = book.imageUrl.split("/images/")[1];
+            fs.unlink(`images/${filename}`, () => {
+                Book.deleteOne({_id: req.params.id})
+                .then(() => resp.status(200).json({ message: "Book successfully deleted !!!"}))
+                .catch(error => resp.status(401).json({error}))
+            })
+        }
+    })
+    .catch(error => resp.status(500).json({error}))
 };
 
 exports.selectBook = (req,resp,next) => {
